add git support for checkcopyrights
