+++ mx.py
@@ -12275,11 +12275,11 @@ def checkcopyrights(args):
             return ArgumentParser.format_help(self) + self._get_program_help()
 
         def _get_program_help(self):
-            help_output = subprocess.check_output([get_jdk().java, '-cp', _cygpathU2W(binDir), 'CheckCopyright', '--help'])
+            help_output = subprocess.check_output([get_jdk().java, '-cp', classpath('com.oracle.mxtool.checkcopy'), 'com.oracle.mxtool.checkcopy.CheckCopyright', '--help'])
             return '\nother argumemnts preceded with --\n' +  help_output
 
     # ensure compiled form of code is up to date
-    myDir, binDir = _compile_mx_class('CheckCopyright')
+    build(['--no-daemon', '--dependencies', 'com.oracle.mxtool.checkcopy'])
 
     parser = CP(prog='mx checkcopyrights')
 
@@ -12298,7 +12298,7 @@ def checkcopyrights(args):
         custom_args = []
         if exists(custom_copyrights):
             custom_args = ['--custom-copyright-dir', custom_copyrights]
-        rc = run([get_jdk().java, '-cp', _cygpathU2W(binDir), 'CheckCopyright', '--copyright-dir', _cygpathU2W(myDir)] + custom_args + args.remainder, cwd=s.dir, nonZeroIsFatal=False)
+        rc = run([get_jdk().java, '-cp', classpath('com.oracle.mxtool.checkcopy'), 'com.oracle.mxtool.checkcopy.CheckCopyright', '--copyright-dir', _mx_home] + custom_args + args.remainder, cwd=s.dir, nonZeroIsFatal=False)
         result = result if rc == 0 else rc
     return result
 
@@ -12981,9 +12981,6 @@ def main():
                 abort('Command timed out after ' + str(_opts.timeout) + ' seconds: ' + ' '.join(commandAndArgs))
             signal.signal(signal.SIGALRM, alarm_handler)
             signal.alarm(_opts.timeout)
-        # compile java resources ahead of time to avoid races in parallel tasks
-        for java_resource in ['CheckCopyright']:
-            _compile_mx_class(java_resource)
         retcode = c(command_args)
         if retcode is not None and retcode != 0:
             abort(retcode)
@@ -12991,7 +12988,7 @@ def main():
         # no need to show the stack trace when the user presses CTRL-C
         abort(1)
 
-version = VersionSpec("5.15.4")
+version = VersionSpec("5.16.1")
 
 currentUmask = None
